Remove --port from docs, moved to package.json
While --port is pretty standard for overriding the port, it becomes a bit non-trivial when it is also defined in the package json. It is also ambiguous since many services can be started with the run command. The advised solution would be to modify the package.json if you need a different port.